commiting to resolve a bunch of the pr comments.
